+++ converter/src/main/kotlin/de/jensklingenberg/htmltocfw/converter/node/CFWComment.kt
@@ -1,8 +1,13 @@
 package de.jensklingenberg.htmltocfw.converter.node
 
+import de.jensklingenberg.htmltocfw.converter.visitor.Visitor
 import org.jsoup.nodes.Comment
 
 class CFWComment(private val comment: Comment) : MyNode {
+    override fun accept(visitor: Visitor) {
+        visitor.visitComment(this)
+    }
+
     override fun toString(): String {
         return ("//" + comment.data + "\n")
     }