+++ converter/src/main/kotlin/de/jensklingenberg/htmltocfw/converter/node/LabelNode.kt
@@ -1,7 +1,8 @@
 package de.jensklingenberg.htmltocfw.converter.node
 
 import de.jensklingenberg.htmltocfw.converter.parseAttributes
-import org.jsoup.nodes.Attributes
+import de.jensklingenberg.htmltocfw.converter.visitor.Visitor
+import org.jsoup.nodes.Element
 import org.jsoup.nodes.Node
 
 /**
@@ -9,36 +10,53 @@ import org.jsoup.nodes.Node
  * [org.jetbrains.compose.web.dom.Label]
  */
 
-class LabelNode(private val htmlAttributes: Attributes, private val childNodes: List<Node>) : MyNode {
+class LabelNode(
+    private val attrs: List<ComposeAttribute>,
+    private val childNodes: List<Node>,
+    val forIdValue: String = ""
+) :
+    MyNode {
     private val ATTR_FOR = "for"
     private val TAG = "Label"
+    override fun accept(visitor: Visitor) {
+        visitor.visitLabel(this)
+    }
 
     override fun toString(): String {
-        var str = "$TAG ("
-        val forValue = htmlAttributes.get(ATTR_FOR)
-        htmlAttributes.remove(ATTR_FOR)
+        var str = "$TAG "
 
-        val attrText = parseAttributes(htmlAttributes.asList())
-        str += attrText
-
-        if (forValue.isNotBlank()) {
-            if (attrText.isNotBlank()) {
-                str += (", ")
-            }
+        val arguments = mutableListOf<String>()
+        if (attrs.isNotEmpty()) {
+            arguments.add(parseAttributes(attrs))
         }
-        str += ("forId = \"${forValue}\"") + ")"
+        arguments.add("forId = \"${forIdValue}\"")
+
+        str += "(" + arguments.joinToString { it } + ")"
+
+        str += "{"
 
         val childNodesText = childNodes.joinToString("") {
             getMyNode(it).toString()
         }
-
-        str += "{"
         if (childNodesText.isNotBlank()) {
             str += "\n" + childNodesText
         }
 
         str += "}\n"
         return str
+    }
+
+    companion object {
+        private val ATTR_FOR = "for"
+
+        fun createLabel(node: Element): LabelNode {
+            val htmlAttributes = node.attributes()
+            val forValue = htmlAttributes.get(ATTR_FOR)
+            htmlAttributes.remove(ATTR_FOR)
+
+            val attrs = node.attributes().map { ComposeAttribute(it.key, it.value) }
+            return LabelNode(attrs, node.childNodes(), forValue)
+        }
     }
 
 }