+++ converter/src/main/kotlin/de/jensklingenberg/htmltocfw/converter/node/TitleNode.kt
@@ -1,6 +1,11 @@
 package de.jensklingenberg.htmltocfw.converter.node
 
+import de.jensklingenberg.htmltocfw.converter.visitor.Visitor
+
 class TitleNode(private val titleText: String) : MyNode {
+    override fun accept(visitor: Visitor) {
+        visitor.visitTitle(this)
+    }
 
     override fun toString(): String {
         return "document.title=\"${titleText}\"\n"